# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF656 - Service Problem Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Problem Management API goal is to provide\
    \ the ability to manage problems in the Service domain. \n\n### Resource \n- serviceProblem\n\
    \n### Operations\nService Problem Management API performs the following operations\
    \ on the resource :\n- Retrieve an entity or a collection of entities depending\
    \ on filter criteria\n- Partial update of an entity (including updating rules)\n\
    - Create an entity (including default values and creation rules)\n- Delete an\
    \ entity (for administration purposes)\n- Manage notification of events"
  version: "4.0.0"
  title: "Service Problem Management"
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
  /serviceProblem:
    get:
      tags:
      - "serviceProblem"
      summary: "List or find ServiceProblem objects"
      description: "This operation list or find ServiceProblem entities"
      operationId: "listServiceProblem"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceProblem"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceProblem"
    post:
      tags:
      - "serviceProblem"
      summary: "Creates a ServiceProblem"
      description: "This operation creates a ServiceProblem entity."
      operationId: "createServiceProblem"
      parameters:
      - in: "body"
        name: "serviceProblem"
        description: "The ServiceProblem to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceProblem_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceProblem"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceProblem"
  /serviceProblem/{id}:
    get:
      tags:
      - "serviceProblem"
      summary: "Retrieves a ServiceProblem by ID"
      description: "This operation retrieves a ServiceProblem entity. Attribute selection\
        \ is enabled for all first level attributes."
      operationId: "retrieveServiceProblem"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceProblem"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceProblem"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceProblem"
    delete:
      tags:
      - "serviceProblem"
      summary: "Deletes a ServiceProblem"
      description: "This operation deletes a ServiceProblem entity."
      operationId: "deleteServiceProblem"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceProblem"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceProblem"
    patch:
      tags:
      - "serviceProblem"
      summary: "Updates partially a ServiceProblem"
      description: "This operation updates partially a ServiceProblem entity."
      operationId: "patchServiceProblem"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceProblem"
        required: true
        type: "string"
      - in: "body"
        name: "serviceProblem"
        description: "The ServiceProblem to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceProblem_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceProblem"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceProblem"
...
definitions:
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"